Extract geocoding into a separate folder
