+++ ingest/geocoding/overpass/README.md
@@ -49,5 +49,5 @@ When numbered addresses are available but Google geocoding is not used, Overpass
 
 ## Related Documentation
 
-- [Geocoding Overview](geocoding-overview.md) - Multi-service geocoding architecture
-- [Message Ingest Pipeline](../../ingest/messageIngest/README.md) - Integration with extraction and conversion stages
+- [Geocoding Router](../README.md)
+- [Message Ingest Pipeline](../../messageIngest/README.md)
