+++ ingest/geocoding/shared/coordinate-utils.ts
